Refactor benchmarks
