+++ tree_performance/benchmarks/c_sequential.py
@@ -1,0 +1,9 @@
+import util
+
+VECTORISED = False
+
+
+def run(ts_path, max_sites):
+    return util.benchmark_external(
+        f"compiled_implementations/c_sequential", ts_path, max_sites
+    )
