+++ tree_performance/benchmarks/cpp_sequential.py
@@ -1,0 +1,9 @@
+import util
+
+VECTORISED = False
+
+
+def run(ts_path, max_sites):
+    return util.benchmark_external(
+        f"compiled_implementations/cpp_sequential", ts_path, max_sites
+    )
